Enum discord::GetMessages

pub enum GetMessages {
    MostRecent,
    Before(MessageId),
    After(MessageId),
    Around(MessageId),
}

Argument to get_messages to specify the desired message retrieval.

Variants
MostRecent

Get the N most recent messages.

Before(MessageId)

Get the first N messages before the specified message.

After(MessageId)

Get the first N messages after the specified message.

Around(MessageId)

Get N/2 messages before, N/2 messages after, and the specified message.