Module discord::model

Struct and enum definitions of values in the Discord model.

Re-exports

pub use self::permissions::Permissions;

Modules

permissions

Structs

ApplicationInfo
Information about the current application and the owner.

Attachment
File upload attached to a message

Ban
A banning of a user

Call
An active group or private call

ChannelId
An identifier for a Channel

ChannelOverride
A channel-specific notification settings override

CurrentUser
Information about the logged-in user

Emoji
A parter custom emoji

EmojiId
An identifier for an Emoji

FriendSourceFlags
Flags for who may add this user as a friend.

Game
Information about a game being played

Group
A group channel, potentially including other users, separate from a server.

Incident
An incident retrieved from the Discord status page.

IncidentUpdate
An update to an incident from the Discord status page. This will typically state what new information has been discovered about an incident.

Invite
Information about an invite

LiveServer
Live server information

Maintenance
Discord status maintenance message.

Member
Information about a member of a server

Mention
A mention targeted at a specific user, channel, or other entity.

Message
Message transmitted over a text channel

MessageId
An identifier for a Message

PermissionOverwrite
A channel-specific permission overwrite for a role or member.

Presence
A members's online status

PrivateChannel
Private text channel to another user

PublicChannel
Public voice or text channel within a server

Reaction

ReadState
Summary of messages since last login

ReadyEvent
The "Ready" event, containing initial state

Relationship
Information on a friendship relationship this user has with another.

RichInvite
Detailed information about an invite, available to server managers

Role
Information about a role

RoleId
An identifier for a Role

Server
Static information about a server

ServerId
An identifier for a Server

ServerInfo
The basic information about a server only

ServerPrune
Representation of the number of member that would be pruned by a server prune operation.

Tutorial
Progress through the Discord tutorial

User
Broadly-applicable user information

UserId
An identifier for a User

UserServerSettings
User settings which influence per-server notification behavior

UserSettings
User settings usually used to influence client behavior

VoiceRegion
Information about an available voice region

VoiceState
A member's state within a voice channel

Enums

Channel
A private or public channel

ChannelType
The type of a channel

Event
Event received over a websocket connection

GameType
A type of game being played.

MessageType
The type of a message

NotificationLevel
Notification level for a channel or server

OnlineStatus
A user's online presence status

PermissionOverwriteType
The type of edit being made to a Channel's permissions.

PossibleServer
A server which may be unavailable

ReactionEmoji
Emoji information sent only from reaction events

RelationshipType
A type of relationship this user has with another.

VerificationLevel
A condition that new users must satisfy before posting in text channels