Enum discord::model::VerificationLevel

pub enum VerificationLevel {
    None,
    Low,
    Medium,
    High,
}

A condition that new users must satisfy before posting in text channels

Variants
None

No verification is needed

Low

Must have a verified email on their Discord account

Medium

Must also be registered on Discord for longer than 5 minutes

High

Must also be a member of this server for longer than 10 minutes

Methods

impl VerificationLevel

fn num(&self) -> u64

fn from_num(num: u64) -> Option<Self>