Module discord::model::permissions

Structs

Permissions
Set of permissions assignable to a Role or PermissionOverwrite

Constants

ADD_REACTIONS
Add emoji reactions to messages

ADMINISTRATOR
Grant all permissions, bypassing channel-specific permissions

ATTACH_FILES

BAN_MEMBERS

CHANGE_NICKNAMES
Change their own nickname

CREATE_INVITE

EMBED_LINKS

EXTERNAL_EMOJIS
Use emojis from other servers

KICK_MEMBERS

MANAGE_CHANNELS
Create channels or edit existing ones

MANAGE_EMOJIS
Manage the emojis in a a server.

MANAGE_MESSAGES
Delete messages by other users

MANAGE_NICKNAMES
Change the nickname of other users

MANAGE_ROLES
Modify roles below their own

MANAGE_SERVER
Change the server's name or move regions

MANAGE_WEBHOOKS
Manage channel webhooks

MENTION_EVERYONE
Trigger a push notification for an entire channel with "@everyone"

READ_HISTORY

READ_MESSAGES

SEND_MESSAGES

SEND_TTS_MESSAGES
Send text-to-speech messages to those focused on the channel

VOICE_CONNECT

VOICE_DEAFEN_MEMBERS

VOICE_MOVE_MEMBERS
Move users out of this channel into another

VOICE_MUTE_MEMBERS

VOICE_SPEAK

VOICE_USE_VAD
When denied, members must use push-to-talk