Struct discord::model::Group

pub struct Group {
    pub channel_id: ChannelId,
    pub icon: Option<String>,
    pub last_message_id: Option<MessageId>,
    pub last_pin_timestamp: Option<String>,
    pub name: Option<String>,
    pub owner_id: UserId,
    pub recipients: Vec<User>,
}

A group channel, potentially including other users, separate from a server.

Fields
channel_id: ChannelId icon: Option<String> last_message_id: Option<MessageId> last_pin_timestamp: Option<String> name: Option<String> owner_id: UserId recipients: Vec<User>
Methods

impl Group

fn decode(value: Value) -> Result<Group>

fn name(&self) -> Cow<str>

Get this group's name, building a default if needed

fn icon_url(&self) -> Option<String>

Returns the formatted URL of the group's icon.

Returns None if the group does not have an icon.