Struct discord::model::IncidentUpdate

pub struct IncidentUpdate {
    pub body: String,
    pub id: String,
    pub incident_id: String,
    pub status: String,
    pub affected_components: Vec<Value>,
    pub created_at: String,
    pub display_at: String,
    pub updated_at: String,
}

An update to an incident from the Discord status page. This will typically state what new information has been discovered about an incident.

Fields
body: String id: String incident_id: String status: String affected_components: Vec<Value> created_at: String display_at: String updated_at: String
Methods

impl IncidentUpdate

pub fn decode(value: Value) -> Result<Self>