Struct discord::model::LiveServer

pub struct LiveServer {
    pub id: ServerId,
    pub name: String,
    pub owner_id: UserId,
    pub voice_states: Vec<VoiceState>,
    pub roles: Vec<Role>,
    pub region: String,
    pub presences: Vec<Presence>,
    pub member_count: u64,
    pub members: Vec<Member>,
    pub joined_at: String,
    pub icon: Option<String>,
    pub large: bool,
    pub channels: Vec<PublicChannel>,
    pub afk_timeout: u64,
    pub afk_channel_id: Option<ChannelId>,
    pub verification_level: VerificationLevel,
    pub emojis: Vec<Emoji>,
    pub features: Vec<String>,
    pub splash: Option<String>,
    pub default_message_notifications: u64,
    pub mfa_level: u64,
}

Live server information

Fields
id: ServerId name: String owner_id: UserId voice_states: Vec<VoiceState> roles: Vec<Role> region: String presences: Vec<Presence> member_count: u64 members: Vec<Member> joined_at: String icon: Option<String> large: bool channels: Vec<PublicChannel> afk_timeout: u64 afk_channel_id: Option<ChannelId> verification_level: VerificationLevel emojis: Vec<Emoji> features: Vec<String> splash: Option<String> default_message_notifications: u64 mfa_level: u64
Methods

impl LiveServer

fn decode(value: Value) -> Result<LiveServer>

fn icon_url(&self) -> Option<String>

Returns the formatted URL of the server's icon.

Returns None if the server does not have an icon.

fn permissions_for(&self, channel: ChannelId, user: UserId) -> Permissions

Calculate the effective permissions for a specific user in a specific channel on this server.