Struct discord::model::Mention

pub struct Mention {
    // some fields omitted
}

A mention targeted at a specific user, channel, or other entity.

A mention can be constructed by calling .mention() on a mentionable item or an ID type which refers to it, and can be formatted into a string using the format! macro:

let message = format!("Hey, {}, ping!", user.mention());

If a String is required, call mention.to_string().